let's do something with that request
This adds a new function, handle_connection, which reads the incoming stream prints it to stdout.

We bring a few new libraries into scope:
    `std::io::prelude::*` has traits for maniplating the stream
    `std::net::TCPStream` is the type passed to our new handle function

Our new handle_connection function uses a mutable stream parameter because the TcpStream instance may manipulate its own internal state. Thus, even though we only read from it, we still need to mark it as mutable.

In order to read from the stream we build a 512 byte buffer on the stack. This simple buffer works fine for common requests, and later we can modify this to accommodate larger or even arbitrarily-sized requests. The buffer is first passed to `stream.read`, where it is loaded from the stream, and then converted to a string using a lossy strategy. The lossy method will replace invalid sequences with the `U+FFFD REPLACEMENT CHARACTER`.