Write a response
This change adds comments documented HTTP requests and responses. We create a basic response as a string and pass it back through `stream.write` and use `stream.flush` to ensure the program waits for the bytes to actually write out before continuing.

We also remove the logging. If everything is working right you'll see a well formed response in your browser.