A brief detour: file handling
Following along with the rust lang book 1.30 tutorial, we take a diversion to implement static file serving.

This change introduces a new static file in the project root. We pull `std::fs::file` into scope to read the file into a string. The `format!` macro formats the response to include this.

This response will be sent regardless of the verb, resource, query-string, or any other details of the response. Next we'll get into validating the response.