Begin the bog
This change starts the transformation from a general server into a 
tarpit.

This removes all code related to reading the stream. We don't really 
care what was requested -- just want to bog down the response.

To that end we also remove all files and templates used to respond with 
specific resources, as well as pruning out `std::fs::File`.

Instead we add an infinite loop that rapidly sends the same HTTP header 
on repeat.

This server responds until the Pipe Breaks. Firefox times this out at 
around 500ms. Since there is no error handling, this panic crashes the 
server.