Introduce Threadpooling
This change add a `lib.rs` namespace that implements threadpooling. 
Threadpooling is used to control how many threads the server will 
consume when running.

Thread pooling requires some gymnastics to implement. This one follows 
the Rust Lang Book v1.30.0 directly, implementing a multithreaded server 
as described in Chapter 20.

In a future version of rust it will be possible to implement an even 
more efficient version of this with async and zero-cost futures. That 
work will live in a diferent branch until stable has access to these 
features.