added command line args
this change uses structop and clap to take a command line argument for the port. these conveniently generate a useful help file and argument handlers. This also introduces a new default behavior whereby the binary will ask the OS to supply a port if none is provided by the user.